requirements.txt
pillow
requests
beautifulsoup4
comics
spidev
gpiozero
pigpio
lgpio
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
flask
feedparser